requirements.txt
langchain
openai<1.0.0
python-dotenv
